refactor: move compact_helpers to src/agent/ alongside compaction logic
