+++ src/agent/compact_helpers.rs
